requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
deprecation==2.1.0
distlib==0.4.0
Django==5.2.6
django-allauth==65.11.2
filelock==3.19.1
flatbuffers==25.9.23
gast==0.6.0
google-auth==2.40.3
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
keras==3.11.3
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
multidict==6.7.0
namex==0.1.0
numpy==2.3.4
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==12.0.0
platformdirs==4.4.0
postgrest==2.22.0
propcache==0.4.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
realtime==2.22.0
requests==2.32.5
rich==14.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==2.0.0
yarl==1.22.0
